first refacto: splitted recipe into : api, recipes, tools, log
