+++ Dockerfile
@@ -1,7 +1,7 @@
 FROM python:2
 ARG proxy
 
-RUN pip install `echo $proxy | sed 's/\(\S\S*\)/--proxy \1/'` simplejson Flask flask_restplus PyYAML nltk elasticsearch pandas Werkzeug scikit-learn[alldeps] geopy jellyfish feather-format tables python-igraph
+RUN pip install `echo $proxy | sed 's/\(\S\S*\)/--proxy \1/'` simplejson Flask flask_restplus PyYAML nltk elasticsearch pandas Werkzeug scikit-learn[alldeps] geopy jellyfish feather-format tables networkx
  
 RUN mkdir -p /matchid/code /matchid/conf/run /matchid/log /matchid/referential_data /data/matchID_test/
 
@@ -16,7 +16,7 @@ VOLUME /matchid/models
 
 EXPOSE 8081
 
-CMD ["./code/recipe.py"]
+CMD ["./code/api.py"]
